requirements.txt
absl-py==2.0.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==20.3.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cryptography==41.0.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2023.12.1
future==1.0.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.29.3
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
jellyfish==0.11.2
Jinja2==3.1.3
joblib==1.3.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.14.0
kiwisolver==1.4.5
lib50==3.0.6
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.25.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.0
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.5.0
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.3.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
RNBIP==1.6.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
submit50==3.1.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.15.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.35.2
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.13
Werkzeug==3.0.1
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
zipp==3.17.0
